requirements.txt
Flask>=2.2.5,<2.3; python_version < "3.8"
Flask>=2.3,<4; python_version >= "3.8"
chinesecalendar>=1.8,<2
pywebview>=4.4,<5; python_version < "3.8"
pywebview>=5,<7; python_version >= "3.8"
